+++ common/src/main/java/com/cobblemon/mod/common/mixin/entity/EntityMixin.java
@@ -14,6 +14,10 @@ import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import com.cobblemon.mod.common.pokemon.Pokemon;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
+import com.llamalad7.mixinextras.sugar.Share;
+import com.llamalad7.mixinextras.sugar.ref.LocalBooleanRef;
+import com.llamalad7.mixinextras.sugar.ref.LocalFloatRef;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -91,68 +95,60 @@ public abstract class EntityMixin {
     }
 
 
-    @Inject(
+    @WrapOperation(
             method = "push(Lnet/minecraft/world/entity/Entity;)V",
-            at = @At("HEAD"),
-            cancellable = true
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;push(DDD)V", ordinal = 0)
     )
-    public void cobblemon$customPushLogic(Entity pushee, CallbackInfo ci) {
-        var pusher = (Entity)(Object)this;
-        var pusherIsPlayerOrPokemon = (pusher instanceof PokemonEntity || pusher instanceof Player);
-        var pusheeIsPlayerOrPokemon = (pushee instanceof PokemonEntity || pushee instanceof Player);
+    public void cobblemon$customPusherLogic(Entity pusher, double x, double y, double z,
+                                            Operation<Void> original,
+                                            @Local(argsOnly = true) Entity pushee,
+                                            @Local(name = "e") double e, @Local(name = "d") double d,
+                                            @Share("pusherIsPlayerOrPokemon") LocalBooleanRef pusherIsPlayerOrPokemon,
+                                            @Share("pusheeIsPlayerOrPokemon") LocalBooleanRef pusheeIsPlayerOrPokemon,
+                                            @Share("pusheeReceivedForce")LocalFloatRef pusheeReceivedForce) {
+        pusherIsPlayerOrPokemon.set(pusher instanceof PokemonEntity || pusher instanceof Player);
+        pusheeIsPlayerOrPokemon.set(pushee instanceof PokemonEntity || pushee instanceof Player);
 
-        // This check prevents doing any calculations for normal or modded mob interactions. Only pokemon will use this.
-        if (pusherIsPlayerOrPokemon && pusheeIsPlayerOrPokemon){
-            if (!pusher.isPassengerOfSameVehicle(pushee)) {
-                if (!pushee.noPhysics && !pusher.noPhysics) {
-                    double d = pushee.getX() - pusher.getX();
-                    double e = pushee.getZ() - pusher.getZ();
-                    double f = Mth.absMax(d, e);
-                    if (f >= (double) 0.01F) {
-                        f = Math.sqrt(f);
-                        d /= f;
-                        e /= f;
-                        double g = (double) 1.0F / f;
-                        if (g > (double) 1.0F) {
-                            g = (double) 1.0F;
-                        }
-
-                        d *= g;
-                        e *= g;
-                        d *= (double) 0.05F;
-                        e *= (double) 0.05F;
-
-                        /**********************************************
-                         * Custom portion of the code:
-                         * Handle collisions with weight in mind. Use
-                         * the weight value of pokemon to determine
-                         * which side gets what portion of the total
-                         * force.
-`                        *********************************************/
-                        // Weights in hectograms
-                        // Check to see if the pusher or pushee is a player. 200ish pounds. Steve's like a 6'6 miner
-                        var pusherWeight = pusher instanceof PokemonEntity pusherPokemon ? pusherPokemon.getPokemon().getSpecies().getWeight() : 900F;
-                        var pusheeWeight = pushee instanceof PokemonEntity pusheePokemon ? pusheePokemon.getPokemon().getSpecies().getWeight() : 900F;
-
-                        // Calculate the portion of force received for both entities. Bigger receives less from the
-                        // collision than the smaller one.
-                        var totalWeight = pusherWeight + pusheeWeight;
-                        var pusherReceivedForce = (pusheeWeight / totalWeight) * 2F;
-                        var pusheeReceivedForce = (pusherWeight / totalWeight) * 2F;
-
-                        // Do the pushing for both entities
-                        if (!pusher.isVehicle() && pusher.isPushable()) {
-                            pusher.push(-d * pusherReceivedForce, (double) 0.0F, -e * pusherReceivedForce);
-                        }
-                        if (!pushee.isVehicle() && pushee.isPushable()) {
-                            pushee.push(d * pusheeReceivedForce, (double) 0.0F, e * pusheeReceivedForce);
-                        }
-                    }
-
-                }
-            }
-            // Do not continue on to the base logic
-            ci.cancel();
+        if (!pusheeIsPlayerOrPokemon.get() || !pusherIsPlayerOrPokemon.get()) {
+            original.call(pusher, x, y, z);
+            return;
         }
+
+        /**********************************************
+         * Custom portion of the code:
+         * Handle collisions with weight in mind. Use
+         * the weight value of pokemon to determine
+         * which side gets what portion of the total
+         * force.
+         *********************************************/
+        // Weights in hectograms
+        // Check to see if the pusher or pushee is a player. 200ish pounds. Steve's like a 6'6 miner
+        var pusherWeight = pusher instanceof PokemonEntity pusherPokemon ? pusherPokemon.getPokemon().getSpecies().getWeight() : 900F;
+        var pusheeWeight = pushee instanceof PokemonEntity pusheePokemon ? pusheePokemon.getPokemon().getSpecies().getWeight() : 900F;
+
+        // Calculate the portion of force received for both entities. Bigger receives less from the
+        // collision than the smaller one.
+        var totalWeight = pusherWeight + pusheeWeight;
+        var pusherReceivedForce = pusheeWeight / totalWeight * 2F;
+        pusheeReceivedForce.set((pusherWeight / totalWeight) * 2F);
+
+        original.call(pusher, -d * pusherReceivedForce, 0.0, -e * pusherReceivedForce);
+    }
+
+    @WrapOperation(
+            method = "push(Lnet/minecraft/world/entity/Entity;)V",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;push(DDD)V", ordinal = 1)
+    )
+    public void cobblemon$customPusheeLogic(Entity pusher, double x, double y, double z, Operation<Void> original,
+                                            @Local(name = "e") double e, @Local(name = "d") double d,
+                                            @Share("pusherIsPlayerOrPokemon") LocalBooleanRef pusherIsPlayerOrPokemon,
+                                            @Share("pusheeIsPlayerOrPokemon") LocalBooleanRef pusheeIsPlayerOrPokemon,
+                                            @Share("pusheeReceivedForce")LocalFloatRef pusheeReceivedForce) {
+        if (!pusheeIsPlayerOrPokemon.get() || !pusherIsPlayerOrPokemon.get()) {
+            original.call(pusher, x, y, z);
+            return;
+        }
+
+        original.call(pusher, d * pusheeReceivedForce.get(), 0.0, e * pusheeReceivedForce.get());
     }
 }